HW2 fix reviewer's comments
