+++ hw02_unpack_string/internal/repeatgroup/repeatgroup.go
@@ -4,7 +4,8 @@ import (
 	"bytes"
 	"fmt"
 	"unicode"
-	"unicode/utf8"
+
+	"github.com/PrideSt/otus-golang/hw02_unpack_string/internal/reader"
 )
 
 // Parser can create Unpacker from string, create internal view from encoded string.
@@ -12,40 +13,10 @@ type Parser interface {
 	ParseString(str string) (Unpacker, error)
 }
 
-// Unpacker can create string from Unpacker, assembly string from internal view.
-type Unpacker interface {
-	Unpack() (string, error)
-}
-
 // RepeatGroup basic struct of internal representation.
 type repeatGroup struct {
 	buffer    []byte
 	repeatCnt int
-}
-
-// GroupStorage contain all groups.
-type GroupStorage struct {
-	rgs []repeatGroup
-}
-
-// Reader is wrapper for utf8 decode function.
-type Reader struct {
-	buffer []byte
-	offset int
-}
-
-// AddRepeatGroup new entry to storage.
-func (gs *GroupStorage) addRepeatGroup(newRG repeatGroup) (newSize int) {
-	gs.rgs = append(gs.rgs, newRG)
-	return len(gs.rgs)
-}
-
-// Add new RepeatGroup entry to storage created it from butes (copy).
-func (gs *GroupStorage) add(b []byte, times int) (newSize int) {
-	chunk := make([]byte, len(b))
-	copy(chunk, b)
-	gs.addRepeatGroup(repeatGroup{chunk, times})
-	return len(gs.rgs)
 }
 
 // as unicode.Digit, but allow only ascii digits [0-9].
@@ -75,21 +46,13 @@ func isJoiner(r rune) bool {
 	return unicode.In(r, unicode.Join_Control)
 }
 
-// flushBuffer create repeatGroup from buffer and cnt and flush buffer.
-func flushBuffer(gs *GroupStorage, buffer *bytes.Buffer, cnt int) {
-	if buffer.Len() > 0 {
-		gs.add(buffer.Bytes(), cnt)
-		buffer.Reset()
-	}
-}
-
 // ParseString converts input string to internal state.
 func ParseString(input string) (Unpacker, error) {
 	var gs GroupStorage
 	var buffer bytes.Buffer
-	reader := Reader{[]byte(input), 0}
+	reader := reader.Make([]byte(input), 0)
 
-	for reader.IsNotEOF() {
+	for !reader.IsEOF() {
 		r, offset := reader.GetNext()
 
 		switch {
@@ -98,13 +61,13 @@ func ParseString(input string) (Unpacker, error) {
 				return GroupStorage{}, fmt.Errorf("invalid format %s, offset: %d, digit can't be the first symbol", input, offset)
 			}
 
-			flushBuffer(&gs, &buffer, int(r-'0'))
+			gs.flushBuffer(&buffer, int(r-'0'))
 			continue
 		case isEscapeSymbol(r):
 			// if escape is last symbol do nothing, add them to buffer like any another
 			// otherwise read next symbol
-			flushBuffer(&gs, &buffer, 1)
-			if reader.IsNotEOF() {
+			gs.flushBuffer(&buffer, 1)
+			if !reader.IsEOF() {
 				// overwrite variables in outer scope
 				r, _ = reader.GetNext()
 			}
@@ -113,7 +76,7 @@ func ParseString(input string) (Unpacker, error) {
 			buffer.WriteRune(r)
 
 			// write next symbol
-			if reader.IsNotEOF() {
+			if !reader.IsEOF() {
 				// overwrite variables in outer scope
 				r, _ = reader.GetNext()
 			} else {
@@ -121,43 +84,11 @@ func ParseString(input string) (Unpacker, error) {
 			}
 		case isSymbolModifier(r) || isMarkNonspacing(r):
 		default:
-			flushBuffer(&gs, &buffer, 1)
+			gs.flushBuffer(&buffer, 1)
 		}
 		buffer.WriteRune(r)
 	}
 
-	flushBuffer(&gs, &buffer, 1)
+	gs.flushBuffer(&buffer, 1)
 	return gs, nil
-}
-
-// Unpack converts internal state to string.
-func (gs GroupStorage) Unpack() (string, error) {
-	var result bytes.Buffer
-	for _, gr := range gs.rgs {
-		if gr.repeatCnt > 0 {
-			newChunk := bytes.Repeat(gr.buffer, gr.repeatCnt)
-			writedLen, err := result.Write(newChunk)
-			if err != nil {
-				return "", fmt.Errorf("unable to write into buffer, %w", err)
-			}
-			if writedLen < len(gr.buffer) {
-				return "", fmt.Errorf("buffer writes partially")
-			}
-		}
-	}
-
-	return result.String(), nil
-}
-
-// GetNext returns next rune from internal buffer.
-func (r *Reader) GetNext() (rune, int) {
-	rn, sz := utf8.DecodeRune(r.buffer[r.offset:])
-	r.offset += sz
-
-	return rn, r.offset
-}
-
-// IsNotEOF returns false when offset goes till end of internal buffer.
-func (r *Reader) IsNotEOF() bool {
-	return r.offset < len(r.buffer)
 }
